Created Methods for Color Sensor and Servos
Do we want to combine all common methods of color and touch sensors into
one?
No. It will be easier for users working in the libraries to call methods
that relate to one sensor  rather than having the position as a
parameter.

Do we want to combine sets and gets of Servos into one method using
array?
Just like the solution to the above question, having the user call
methods relating to only one servo will save time due to not writing the
parameter.